this is good libft

+++ LIBFT/ft_memset.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, unsigned int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
 	char	*ptr;
